Disabled the tests, which try to access internet.
It seems that sequence ontology is needed for the tests, and
there are attempts to donwload them from internet. Not only the
ressource is error 500 today, but anyway the buildds do not give
net access to the packages being built.

Maybe we can package SO and build-depend on it if we want the tests
in the build logs...

http://song.cvs.sourceforge.net/song/ontology/