Disabling call to /usr/bin/gbrowse_metadb_config.
Since it was creating var/cache/gbrowse/sessions, debian/gbrowse.dirs is
modified accordingly to create it at build time.

More in general, /usr/bin/gbrowse_metadb_config is creating and removing
directories in /usr and /var.  I do not think that this should be done in
a maintainer script.  Since it is not essential for runnign the demos,
I disabled it for the moment.