This is a update of the Qiime 1.2.1 package to build Qiime 1.3.
Unfortunately, this is now a fork from the Debian trunk because current versions
of Debian/Ubuntu now build Python with "dh --with python2" but I can't use this for
backports to Lucid Lynx.

Anyone wanting to package Qiime 1.3 or above for Deb Med should ignore this fork, and
work from the much neater trunk.  However, you should take a look at my last changelog
entry as there are some important changes from the packaging point of view - primarily
the inclusion of a small piece of compiled code and the incorporation of Denoiser within
the core Qiime codebase.