Ivo uses Matlab with SBML in his daily routine. There is no way he can maintain the package without that Matlab link.  Octave is not a substitute, was tested. We once came up with this approach of having the control file adapted to show and build the libsbml5-matlab package when the development files for Matlab are available and not if not. This way, everything works fine on all architectures, with (amd64 only) Matlab or without.
Mathworks had sent me a license for free only to help with Debian
packaging when I sponsored the initial package. Maybe we should mention
that somewhere and seek some additional permanent way of support through
them.  Matlab is surprisingly important to many in system biology, so
we should not ignore it. I have removed the other uploaders of libsbml5
for the moment. Those wishing to collaborate with Ivo on the maintenance
of the SBML packages please have a chat with him, first.

Cheers,

Steffen