I've packaged the latest hmmer for Bio-Linux/Ubuntu LTS.
Note that despite having a 'b' in the version it is on the website as a stable
release.  I've modified the watchfile to spot it.

I switched from CDBS to DH as I can't get on with CDBS and modifications
were required to make the build happy.  Apologies to the
maintainers if you don't like this but if you don't like it you can always revert
(to r14447).