ATTENTION: Tim, please test!!!
I replaced a code copy of the package libjs-jquery-ui in a newer version via

  replace        usr/share/javascript/jquery-ui/ui/jquery-ui.custom.min.js       usr/share/emperor/support_files/js/jquery-ui-1.8.17.custom.min.js

which is *hackish* since it encodes a wrong version in a file name (Debian has version 1.10).  The alternative would be to patch format.py but before we go on doing things like this we should verify whether this approach wo replace JS libraries by higher versions might really work!